TwitterStreaming/requirement.txt
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-language==1.2.0
googleapis-common-protos==1.5.9
grpcio==1.20.0
idna==2.8
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kafka==1.3.5
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
nltk==3.4
notebook==5.7.8
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
py4j==0.10.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
Pygments==2.3.1
pyparsing==2.4.0
pyrsistent==0.14.11
PySocks==1.6.8
pyspark==2.4.1
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.12.0
terminado==0.8.2
testpath==0.4.2
textblob==0.15.3
tornado==6.0.2
traitlets==4.3.2
tweepy==3.7.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
